+++ src/pyxtp/binds/main.cc
@@ -5,4 +5,6 @@ int main() {
   pyxtp::PyXTP pyxtp;
   std::cout << "starting:\n";
   pyxtp.Initialize("eanalyze", 1);
+
+  pyxtp::call_calculator("eanalyze", 1);
 }
